+++ src/main/model.ts
@@ -45,6 +45,7 @@ export type ModelDataset = Dataset | RawDataDataset;
 export type GetDecodeParams = (pageId: string) => (entries: [string, any][]) => object;
 
 export interface SetDataOptions {
+    setPrepareStepTitle?: (name: string) => Promise<void>;
     dataset?: Dataset;
 }
 
@@ -64,10 +65,12 @@ export type PrepareContextApiWrapper = {
     contextApi: PrepareContextApi | LegacyPrepareContextApi;
 };
 export interface PrepareContextApi {
+    setStageTitle: (name: string) => Promise<void>;
     rejectData: (message: string, extra: any) => void;
     markers: Record<string, (value: unknown) => void>;
 }
 export interface LegacyPrepareContextApi {
+    setStageTitle: (name: string) => Promise<void>;
     rejectData: (message: string, extra: any) => void;
     defineObjectMarker<T extends object>(name: string, options: ObjectMarkerConfig<T>): ObjectMarker<T>['mark'];
     lookupObjectMarker(value: any, type?: string): ObjectMarkerDescriptor<object> | null;
@@ -227,7 +230,7 @@ export class Model<
         this.prepare = fn;
     }
 
-    setData(data: unknown, options: SetDataOptions) {
+    setData(data: unknown, options?: SetDataOptions) {
         options = options || {};
 
         // mark as last setData promise
@@ -243,8 +246,8 @@ export class Model<
         };
 
         const prepareApi = this.#legacyPrepare
-            ? createLegacyExtensionApi(this)
-            : createExtensionApi(this);
+            ? createLegacyExtensionApi(this, options)
+            : createExtensionApi(this, options);
         const setDataPromise = Promise.resolve()
             .then(() => {
                 checkIsNotPrevented();